Rebalancing due to miscalculations
The IE config file says that the loss ratio for wire is applied every 16 blocks,
but this is incorrect. According to the code, the penalty is actually applied
every X blocks, where X is the maximum length of that tier of wire.

Long story short, the loss ratio I selected was based on the incorrect loss calculations,
so I've done a rebalance to match the correct calculations.